+++ mex/editor/search/transform.py
@@ -1,14 +1,14 @@
 from collections.abc import Iterable
 
-from mex.common.models import AnyPreviewModel
+from mex.common.models import AnyExtractedModel, AnyPreviewModel
 from mex.editor.search.models import SearchResult
 from mex.editor.transform import transform_models_to_preview, transform_models_to_title
 
 
 def transform_models_to_results(
-    models: Iterable[AnyPreviewModel],
+    models: Iterable[AnyPreviewModel | AnyExtractedModel],
 ) -> list[SearchResult]:
-    """Convert a list of merged model previews into search result models."""
+    """Convert a list of models into a list of search result models."""
     return [
         SearchResult(
             identifier=model.identifier,