+++ mex/mex.py
@@ -15,6 +15,8 @@ from mex.editor.edit.state import EditState
 from mex.editor.ingest.main import index as ingest_index
 from mex.editor.ingest.state import IngestState
 from mex.editor.login.main import index as login_index
+from mex.editor.merge_items.main import index as merge_items_index
+from mex.editor.merge_items.state import MergeState
 from mex.editor.rules.state import RuleState
 from mex.editor.search.main import index as search_index
 from mex.editor.search.state import SearchState
@@ -37,6 +39,16 @@ app.add_page(
         SearchState.load_search_params,
         SearchState.refresh,
         SearchState.resolve_identifiers,
+    ],
+)
+app.add_page(
+    merge_items_index,
+    route="/merge-items",
+    title="MEx Editor | Merge Items",
+    on_load=[
+        State.check_login,
+        State.load_nav,
+        MergeState.refresh,
     ],
 )
 app.add_page(